requirements.txt
streamlit==1.31.0

langchain==0.1.16
langchain-community==0.0.32
langchain-groq==0.1.3
langchain-text-splitters==0.0.1

chromadb==0.4.24
faiss-cpu==1.7.4

sentence-transformers==2.2.2
huggingface-hub==0.20.3

pypdf==4.0.1
python-dotenv==1.0.1
